+++ UUIDUtil/UUIDv4.cs
@@ -1,6 +1,6 @@
 ﻿// SPDX-License-Identifier: Apache-2.0
 //
-//   Copyright 2021 TensionDev <TensionDev@outlook.com>
+//   Copyright 2021 - 2026 TensionDev <TensionDev@outlook.com>
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.